split selectstackscriptpanel content into different components
